feat(): general brick contracts for providers

+++ packages/build-config-factory/src/generateBrickContracts.js
@@ -2,16 +2,17 @@ const path = require("path");
 const fs = require("fs-extra");
 const globby = require("globby");
 const yaml = require("js-yaml");
+const changeCase = require("change-case");
 const {
   webpack,
   bricks: { webpackContractsFactory },
 } = require("@next-core/webpack-config-factory");
 
-module.exports = function generateBrickContracts(dir) {
+module.exports = function generateBrickContracts(dir, isProviderBricks) {
   const brickEntriesFilePath = path.join(dir, "dist/brick-entries.json");
-  const brickEntries = JSON.parse(
-    fs.readFileSync(brickEntriesFilePath, "utf-8")
-  );
+  const brickEntries = isProviderBricks
+    ? { index: "src/index.ts" }
+    : fs.readJsonSync(brickEntriesFilePath);
 
   webpack(webpackContractsFactory(dir, brickEntries), async (err, stats) => {
     if (err || stats.hasErrors()) {
@@ -26,10 +27,12 @@ module.exports = function generateBrickContracts(dir) {
       // Done processing
       console.log("contracts.log generated.");
 
-      fs.remove(brickEntriesFilePath);
+      if (!isProviderBricks) {
+        fs.remove(brickEntriesFilePath);
+      }
 
-      const version = (await fs.readJson(path.join(dir, "package.json")))
-        .version;
+      const pkg = await fs.readJson(path.join(dir, "package.json"));
+      const pkgLastName = pkg.name.split("/")[1];
       const bricks = (await fs.readJson(path.join(dir, "dist/bricks.json")))
         .bricks;
       const depsMap = new Map();
@@ -42,28 +45,48 @@ module.exports = function generateBrickContracts(dir) {
         contractFiles.map(async (filePath) => {
           const source = await fs.readFile(filePath, "utf-8");
           const contracts = source.match(contractRegExp);
-          const brick = path.basename(filePath, ".js.contracts");
-          depsMap.set(
-            brick,
-            contracts?.map((item) => {
-              const [contract, version] = item.split("@");
-              return {
-                type: "contract",
-                contract,
-                version: version ?? "*",
-              };
-            })
-          );
+          if (isProviderBricks) {
+            if (contracts) {
+              for (const item of contracts) {
+                const [contract, version] = item.split("@");
+                depsMap.set(
+                  `${pkgLastName}.${contract
+                    .split(".")
+                    .slice(-2)
+                    .map((seg) => changeCase.paramCase(seg))
+                    .join("-api-")}`,
+                  {
+                    type: "contract",
+                    contract,
+                    version: version ?? "*",
+                  }
+                );
+              }
+            }
+          } else {
+            const brick = path.basename(filePath, ".js.contracts");
+            depsMap.set(
+              brick,
+              contracts?.map((item) => {
+                const [contract, version] = item.split("@");
+                return {
+                  type: "contract",
+                  contract,
+                  version: version ?? "*",
+                };
+              })
+            );
+          }
         })
       );
       const implementedBricks = bricks.map((brick) => ({
         type: "brick",
         brick,
-        version,
+        version: pkg.version,
         deps: depsMap.get(brick),
       }));
 
-      console.log({ implementedBricks });
+      console.log({ implementedBricks, depsMap });
 
       await fs.writeFile(
         path.join(dir, "deploy/contract.yaml"),
